requirements.txt
customtkinter
pywinstyles
hPyT
pillow